FIX: mobile share canonical url
