+++ app/assets/javascripts/discourse/tests/index.html
@@ -7,8 +7,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta property="og:title" content="Discourse Tests" />
     <meta property="og:url" content="{{rootURL}}" />
-    <meta property="twitter:title" content="Discourse Tests" />
-    <meta property="twitter:url" content="{{rootURL}}" />
+    <meta name="twitter:title" content="Discourse Tests" />
+    <meta name="twitter:url" content="{{rootURL}}" />
 
     <meta name="color-scheme" content="light dark">
